Move CouchTestCase to separate file
